+++ src/app/exercise/exercises/utility/exerciseAttributes/BaseCommonSettingsExercise.ts
