+++ src/app/exercise/exercises/utility/exerciseAttributes/chordProgressionExercise.ts
