+++ src/app/exercise/exercises/utility/exerciseAttributes/defaultSettings.ts
