+++ src/app/exercise/exercises/utility/exerciseAttributes/tonalExercise.ts
